[test][embedded] Declare embedded-synchronization as a test dependency
PR #72293 added the synchronization module to embedded, and it was added
as dependency of `ALL`, but not of the `-test-depends` targets. If one
is using the `-test-depends` targets, the synchronization library for
embedded will not be build, and the test will fail.

This is a follow up to #72329. I had a partial build to test that, and
I did not notice this was also missing.