Sema: Do not diagnose set accessor availability for `InOutExpr`s inside of `LoadExpr`s.
The recent deprecation of the set accessor for `CommandLine.arguments` exposed
a bug in availability checking:

```
 for arg in CommandLine.arguments[1...] {
                      ╰─ warning: setter for 'arguments' is deprecated: ...
```

The parser generates an `InOutExpr` in the AST for the subscript access in the
code above and the availabilty checker was unconditionally diagnosing set
accessors for `InOutExprs`, resulting in spurious availability warnings for
read-only accesses of properties. The fix is to check whether there is an
enclosing `LoadExpr` in the AST and only diagnose the getter accessor in that
case.

Resolves rdar://124566405