[ast] Improve the debug dump for ActorIsolation.
Specifically, I made it so that when one calls dump(), we put in a newline
afterwards so we get a nice easy to read msg in the debugger. I also defined
dumpForDiagnostics() to make it easy to dump out in the debugger how the
ActorIsolation will show up in the diagnostics.

(cherry picked from commit e6a1d83ef82d9d49f0ddcd7f4fc18e7a5b8e83ec)