Push isolated(any) conversions down through multiple levels of conversion.
I'm not sure there's any way to test this right now.  We don't naturally
emit multiple function conversions on a single operand, and the only way
to get a similar effect is to coerce, which interrupts the application of
`@_inheritActorContext`.  So I think this is dead code until we add closure
isolation controls, and even then it might be dead unless we allow coercion
of isolated closures, which maybe we won't.  But it's the right thing to do
in the abstract, and I was thinking of it now.