[region-isolation] Wordsmith "{access,use} here could race".
I am doing this since it isn't always going to be an access. We may not have
memory. We are talking about uses here!

I was able to just use sed so it was an easy fix.

(cherry picked from commit 2f9b5197587d58018a70eb3da55726c90539cc0e)