[concurrency] Diagnose synchronous sendable actor isolated functions that have their actor isolation type erased.
The problem that we are running into here is that:

1. actor isolated functions are considered sendable... so we cannot use region
based isolation here.

2. But, structurally a synchronous actor isolated function can only be called
from a context where we know that it is actor isolated so the caller can hop to
the correct executor. Once the value is type erased and passed to a caller, we
no longer have that information potentially creating races.

Since I can't use region isolation here, I created a type checker check instead.

rdar://118206647