Revert adoption of Int128 in Duration for now
It would prevent using toolchain downloads on macOS 14 (which is
really a bug in how toolchains are configured that should be fixed,
but that's outside the scope of this PR), so restore the old _Int128
type and continue using it as an implementation detail for now.
We can switch over sometime post-6.0 release.

Also underscores the _low and _high properties and the corresponding
initializer, as @lorentey correctly requested.